Component erroring broke Poison.decode
Due to the original return from the process method if there was a
failure it would cause the Poison.decode to kill the process for that request.
Handling it this way (similar to how Requester handles it) means that it
will still be able to parse a empty envelope while letting us know that
the component has timed out or their is a server error.