Different approach to start ConCache OTP App in test.
Previous approach was calling the prod children function as well. This was in my
opinion a bit hidden and could became problematic.

Also it was producing an annoying log in test mode which took me a few mins to
debug. At every run I was getting:
```
{"ttl":null,"message":"Invalid TTL Value, you must provide an integer value for
the tll","level":"error","datetime":"2019-05-15T16:18:22.281000Z"}
```
which was caused by the prod env app complaining for the TTL not being set.